Add mediaquery for large screens
Just fix the auto-fitting columns for the cards. All screens have a minwidth of 300px and 1fr for max; on larger screens it's the other way around, max = 300px.

With other solutions I tried, like setting a max width on one of the containers, the hero triangle gets smaller too.